requirements.txt
# AI/LLM
anthropic==0.34.0

# 音声生成・処理
google-cloud-texttospeech==2.16.3
elevenlabs==1.5.0
pydub==0.25.1

# 動画生成
moviepy==1.0.3
Pillow==10.4.0

# Google APIs
gspread==6.1.2
oauth2client==4.1.3
google-api-python-client==2.147.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# ユーティリティ
requests==2.32.3
python-dotenv==1.0.1
pyyaml==6.0.2
python-dateutil==2.9.0

# デプロイ
gunicorn==23.0.0

# 開発用
pytest==8.3.3
black==24.8.0
flake8==7.1.1